Changed example tree to be 0-indexed.
The "intersect" method in rectanglesOverlap and rectangleInside are INCLUSIVE, meaning (0,8) and (8,16) will count as an "overlap", even though for our QT they should not be.
So I changed the example tree to be 0 index and have leaves be like (0,7) and (8, 15) so that they don't overlap.
However this brings up some issues when a square cannot be split into even squares.
It may not matter since they are rectangles, but all the rectangles SHOULD be the same size, I believe.